refactor: better code

+++ src/data/usecases/user/DbUpdateAvatar.ts
@@ -1,6 +1,7 @@
+import { IUpdateAvatarRepository } from '@/data/protocols/db';
 import { IFindUserByIdRepository } from '@/data/protocols/db/user/IFindUserByIdRepository';
-import { IUpdateAvatarRepository } from '@/data/protocols/db/user/IUpdateAvatarRepository';
-import { IAccount, IStorageProvider, IUpdateAvatar } from './DbUserProtocols';
+import { IUpdateAvatar, UpdateAvatar } from '@/domain/usecases/account/IUpdateAvatar';
+import { IStorageProvider } from './DbUserProtocols';
 
 export class DbUpdateAvatar implements IUpdateAvatar {
   constructor(
@@ -9,21 +10,12 @@ export class DbUpdateAvatar implements IUpdateAvatar {
     private readonly updateAvatarRepository: IUpdateAvatarRepository,
   ) {}
 
-  async updateAvatar(userId:string, avatarFileName: string): Promise<IAccount> {
+  async updateAvatar({ userId, fileName }: UpdateAvatar.Params): Promise<UpdateAvatar.Result> {
     const user = await this.findUserByIdRepository.findById(userId);
-
-    if (!user) {
-      return null;
-    }
-
-    if (user.avatar) {
-      await this.storageProvider.deleteFile(user.avatar);
-    }
-
-    const filename = await this.storageProvider.saveFile(avatarFileName);
-
-    user.avatar = filename;
-    await this.updateAvatarRepository.updateAvatar(userId, filename);
-    return user;
+    if (!user) return null;
+    if (user.avatar) await this.storageProvider.deleteFile(user.avatar);
+    const fileUpadated = await this.storageProvider.saveFile(fileName);
+    const avatar = await this.updateAvatarRepository.updateAvatar({ userId, fileName: fileUpadated });
+    return avatar;
   }
 }
